+++ app/src/main/java/m/derakhshan/done/database/dao/TasksDAO.kt
@@ -20,17 +20,17 @@ interface TasksDAO {
 
     @Query("SELECT * FROM Tasks WHERE id = :id")
     fun getTask(
-        id: Int
+        id: Long
     ): TasksModel
 
     @Query("UPDATE Tasks SET taskName=:taskName WHERE id = :id")
-    fun updateTaskTitle(taskName: String, id: Int)
+    fun updateTaskTitle(taskName: String, id: Long)
 
     @Query("SELECT * FROM Tasks order by status DESC , id ASC")
     fun getAllTasks(): LiveData<List<TasksModel>>
 
     @Query("UPDATE Tasks SET status = :status WHERE id = :id")
-    fun updateStatus(status: TaskStatus, id: Int)
+    fun updateStatus(status: TaskStatus, id: Long)
 
     @Delete
     fun delete(task: TasksModel)
