+++ app/src/main/java/m/derakhshan/done/main/MainViewModel.kt
@@ -12,7 +12,6 @@ import m.derakhshan.done.database.models.TaskStatus
 import m.derakhshan.done.database.models.TasksModel
 import java.util.*
 import javax.inject.Inject
-import kotlin.collections.ArrayList
 import kotlin.collections.HashMap
 
 @HiltViewModel
@@ -51,6 +50,7 @@ class MainViewModel @Inject constructor(private val database: TasksDatabase) : V
             )
         }
     }
+
 
     fun restoreTask(task: TasksModel) {
         viewModelScope.launch(Dispatchers.Default) {
@@ -91,7 +91,7 @@ class MainViewModel @Inject constructor(private val database: TasksDatabase) : V
         this.checkGreeting = false
     }
 
-    fun clearSubTasksOfDeletedTask(items: ArrayList<Int>) {
+    fun clearSubTasksOfDeletedTask(items: ArrayList<Long>) {
         viewModelScope.launch(Dispatchers.Default) {
             for (item in items)
                 database.subTasksDAO.deleteAllSubTasks(item)
